+++ Logitude/Unifreight.Data/AmitalModel/EntityPOCOs/AmitalDataModel.CCUSUPITEM.cs
@@ -254,7 +254,7 @@ namespace Unifreight.Data.AmitalModel.EntityPOCOs
         /// There are no comments for RAISEPERCENT in the schema.
         /// </summary>
         [System.Runtime.Serialization.DataMember]
-        public virtual global::System.Nullable<long> RAISEPERCENT
+        public virtual global::System.Nullable<decimal> RAISEPERCENT
         {
             get
             {
@@ -269,14 +269,14 @@ namespace Unifreight.Data.AmitalModel.EntityPOCOs
                 }
             }
         }
-        private global::System.Nullable<long> _RAISEPERCENT;
+        private global::System.Nullable<decimal> _RAISEPERCENT;
 
     
         /// <summary>
         /// There are no comments for RAISEVALUE in the schema.
         /// </summary>
         [System.Runtime.Serialization.DataMember]
-        public virtual global::System.Nullable<long> RAISEVALUE
+        public virtual global::System.Nullable<decimal> RAISEVALUE
         {
             get
             {
@@ -291,14 +291,14 @@ namespace Unifreight.Data.AmitalModel.EntityPOCOs
                 }
             }
         }
-        private global::System.Nullable<long> _RAISEVALUE;
+        private global::System.Nullable<decimal> _RAISEVALUE;
 
     
         /// <summary>
         /// There are no comments for NISVALUE in the schema.
         /// </summary>
         [System.Runtime.Serialization.DataMember]
-        public virtual global::System.Nullable<long> NISVALUE
+        public virtual global::System.Nullable<decimal> NISVALUE
         {
             get
             {
@@ -313,14 +313,14 @@ namespace Unifreight.Data.AmitalModel.EntityPOCOs
                 }
             }
         }
-        private global::System.Nullable<long> _NISVALUE;
+        private global::System.Nullable<decimal> _NISVALUE;
 
     
         /// <summary>
         /// There are no comments for QUANTITY in the schema.
         /// </summary>
         [System.Runtime.Serialization.DataMember]
-        public virtual global::System.Nullable<long> QUANTITY
+        public virtual global::System.Nullable<decimal> QUANTITY
         {
             get
             {
@@ -335,14 +335,14 @@ namespace Unifreight.Data.AmitalModel.EntityPOCOs
                 }
             }
         }
-        private global::System.Nullable<long> _QUANTITY;
+        private global::System.Nullable<decimal> _QUANTITY;
 
     
         /// <summary>
         /// There are no comments for EXTRAQNTY in the schema.
         /// </summary>
         [System.Runtime.Serialization.DataMember]
-        public virtual global::System.Nullable<long> EXTRAQNTY
+        public virtual global::System.Nullable<decimal> EXTRAQNTY
         {
             get
             {
@@ -357,7 +357,7 @@ namespace Unifreight.Data.AmitalModel.EntityPOCOs
                 }
             }
         }
-        private global::System.Nullable<long> _EXTRAQNTY;
+        private global::System.Nullable<decimal> _EXTRAQNTY;
 
     
         /// <summary>
@@ -386,7 +386,7 @@ namespace Unifreight.Data.AmitalModel.EntityPOCOs
         /// There are no comments for IMPORTADDITION in the schema.
         /// </summary>
         [System.Runtime.Serialization.DataMember]
-        public virtual global::System.Nullable<long> IMPORTADDITION
+        public virtual global::System.Nullable<decimal> IMPORTADDITION
         {
             get
             {
@@ -401,7 +401,7 @@ namespace Unifreight.Data.AmitalModel.EntityPOCOs
                 }
             }
         }
-        private global::System.Nullable<long> _IMPORTADDITION;
+        private global::System.Nullable<decimal> _IMPORTADDITION;
 
     
         /// <summary>
@@ -474,7 +474,7 @@ namespace Unifreight.Data.AmitalModel.EntityPOCOs
         /// There are no comments for AGNTPAYCUST in the schema.
         /// </summary>
         [System.Runtime.Serialization.DataMember]
-        public virtual global::System.Nullable<long> AGNTPAYCUST
+        public virtual global::System.Nullable<decimal> AGNTPAYCUST
         {
             get
             {
@@ -489,14 +489,14 @@ namespace Unifreight.Data.AmitalModel.EntityPOCOs
                 }
             }
         }
-        private global::System.Nullable<long> _AGNTPAYCUST;
+        private global::System.Nullable<decimal> _AGNTPAYCUST;
 
     
         /// <summary>
         /// There are no comments for AGNTPAYTAX in the schema.
         /// </summary>
         [System.Runtime.Serialization.DataMember]
-        public virtual global::System.Nullable<long> AGNTPAYTAX
+        public virtual global::System.Nullable<decimal> AGNTPAYTAX
         {
             get
             {
@@ -511,14 +511,14 @@ namespace Unifreight.Data.AmitalModel.EntityPOCOs
                 }
             }
         }
-        private global::System.Nullable<long> _AGNTPAYTAX;
+        private global::System.Nullable<decimal> _AGNTPAYTAX;
 
     
         /// <summary>
         /// There are no comments for AGNTPAYBITHA in the schema.
         /// </summary>
         [System.Runtime.Serialization.DataMember]
-        public virtual global::System.Nullable<long> AGNTPAYBITHA
+        public virtual global::System.Nullable<decimal> AGNTPAYBITHA
         {
             get
             {
@@ -533,14 +533,14 @@ namespace Unifreight.Data.AmitalModel.EntityPOCOs
                 }
             }
         }
-        private global::System.Nullable<long> _AGNTPAYBITHA;
+        private global::System.Nullable<decimal> _AGNTPAYBITHA;
 
     
         /// <summary>
         /// There are no comments for NIDHEMEHESPCNT in the schema.
         /// </summary>
         [System.Runtime.Serialization.DataMember]
-        public virtual global::System.Nullable<long> NIDHEMEHESPCNT
+        public virtual global::System.Nullable<decimal> NIDHEMEHESPCNT
         {
             get
             {
@@ -555,14 +555,14 @@ namespace Unifreight.Data.AmitalModel.EntityPOCOs
                 }
             }
         }
-        private global::System.Nullable<long> _NIDHEMEHESPCNT;
+        private global::System.Nullable<decimal> _NIDHEMEHESPCNT;
 
     
         /// <summary>
         /// There are no comments for NIDHEMASPCNT in the schema.
         /// </summary>
         [System.Runtime.Serialization.DataMember]
-        public virtual global::System.Nullable<long> NIDHEMASPCNT
+        public virtual global::System.Nullable<decimal> NIDHEMASPCNT
         {
             get
             {
@@ -577,7 +577,7 @@ namespace Unifreight.Data.AmitalModel.EntityPOCOs
                 }
             }
         }
-        private global::System.Nullable<long> _NIDHEMASPCNT;
+        private global::System.Nullable<decimal> _NIDHEMASPCNT;
 
     
         /// <summary>
@@ -1046,7 +1046,7 @@ namespace Unifreight.Data.AmitalModel.EntityPOCOs
         /// There are no comments for STSQNTY in the schema.
         /// </summary>
         [System.Runtime.Serialization.DataMember]
-        public virtual global::System.Nullable<long> STSQNTY
+        public virtual global::System.Nullable<decimal> STSQNTY
         {
             get
             {
@@ -1061,7 +1061,7 @@ namespace Unifreight.Data.AmitalModel.EntityPOCOs
                 }
             }
         }
-        private global::System.Nullable<long> _STSQNTY;
+        private global::System.Nullable<decimal> _STSQNTY;
 
     
         /// <summary>
